// src/app/shared/model/supplierpayment-detail.model.ts
export class SupplierPaymentDetail
{
    userOrderId: string='';
    supplierPhoneNo: string='';
    price: string='';
    orderDate: string='';
    deliveryDate: string='';
    deliveryStatus: string='';
    companyPaymentStatus: string='';
    companyPayment: string='';
    companyPaymentDate: string='';
    supplierPaymentStatus: string='';
    supplierPayment: string='';
    supplierPaymentDate: string='';    
    transactionId: string='';
    comments: string='';
    miscellenous: string='';
}